qcacld-3.0: Don't skip GRO flush for non-legacy clients
Currently GRO flush is being skipped if bus bandwidth level
is low this is resulting in rx packets to be in queue for more
time and affecting over all RTT value.

To avoid this for non legacy clients GRO flush is not being skipped
even at low throughput level.

Change-Id: I5ce64ee2bd5f332cf4db38116b1bbf400220269b
CRs-Fixed: 2978843
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>