qcacld-3.0: Remove concurrent max interfaces flag
Currently the number of netdev interfaces created is equal to
the max number of vdev's supported. There shouldn't be a problem
w.r.t to adding the number of interfaces, the number of interfaces
which can active at a time is already controlled by the object
manager.

Remove the checks required to add the interfaces greater than
the max supported interfaces.

Change-Id: I5f9d1ca343788a77180a76f961fcc511feff4036
CRs-Fixed: 3010712
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>