msm: camera: isp: Associate IFE CDM user data with context
In scheduling delays or race conditions, it has been observed that
cookie value and the request id in userdata returned from CDM
do not match. This can happen due to CDM worker threads getting
delayed. Userdata attached with the CDM is the pointer in the
request. By the time, the cdm callback is received there is a
possibility that the request data has been reset and that request
node being re-used for another request. This can cause issues
particularly in race conditions.
To handle the condition, this commit associates the userdata and
with the context. Also, while submitting the request, the request
id is saved with context, which can be compared during the cdm
callback.

Change-Id: If09bc823602621e7a5e48e8eeb843c4abbeca016
CRs-Fixed: 3015161
Signed-off-by: Gaurav Jindal <gjindal@codeaurora.org>