qcacld-3.0: Do not cache connection info for non sta interface
Currently driver does not check the mode of the current interface
whether it is STA of CLI and it caches the connection info, in this
process it allocates the memory to cache he operation, this memory
get freed only for station type adapter because of which there is
a possibility of memleak for CLI interface.

To avoid above issue, add a change to only cache the information
STA interface.

Change-Id: Ic68d71dfd9887600fe1a287de66d735c2c802491
CRs-Fixed: 3028583
Signed-off-by: Hafidz Muzakky <ais.muzakky@gmail.com>